Updated ToDoManager methods to implement the entites so that it now loads entities and saves entities. Also updated all Id types to be int from long. Also updated the entity factory to contain a method for getting an entity by Id.

+++ ToDoManager/Models/BaseModel.cs
@@ -4,7 +4,7 @@ namespace Models
 {
     public class BaseModel
     {
-        public long? Id { get; set; }
+        public int? Id { get; set; }
         public DateTime DateCreated { get; set; }
         public DateTime? DateUpdated { get; set; }
         public string UpdateUsername { get; set; }
